getting closer... and refactoring
